% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{addDdg.N.frame}
\alias{addDdg.N.frame}
\title{Calculates nucleotide ddG values from nucleotide design and beta values for views of DNA probe.}
\usage{
addDdg.N.frame(data, model, flatCM, window)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}

\item{flatCM}{Vector version of N.values. Built from as.vector() of model@features@N@N.values with rownames and colnames set to NULL.}

\item{window}{Binding view on DNA probe.}
}
\description{
Calculates nucleotide ddG values from nucleotide design and beta values for views of DNA probe.
}
