% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{addDdg.Shape.frame}
\alias{addDdg.Shape.frame}
\title{Adds window-specific shape ddG contributions for each binding window of the probe.}
\usage{
addDdg.Shape.frame(data, model, window)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}

\item{window}{Binding view on DNA probe.}
}
\description{
Adds window-specific shape ddG contributions for each binding window of the probe.
}
