% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDesignMatrix.R
\name{addDesignMatrix.N}
\alias{addDesignMatrix.N}
\title{Adds design matrix for nucleotide features.}
\usage{
addDesignMatrix.N(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Adds design matrix for nucleotide features.
}
