% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDesignMatrix.R
\name{addDesignMatrix.Round}
\alias{addDesignMatrix.Round}
\title{Adds design matrix for round features.}
\usage{
addDesignMatrix.Round(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Adds design matrix for round features.
}
