% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDesignMatrix.R
\name{addDesignMatrix.Shape}
\alias{addDesignMatrix.Shape}
\title{Adds design matrix for shape features.}
\usage{
addDesignMatrix.Shape(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Adds design matrix for shape features.
}
