% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{addNddG}
\alias{addNddG}
\title{Adds nucleotide ddG values for views of DNA probe.}
\usage{
addNddG(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Adds nucleotide ddG values for views of DNA probe.
}
