% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{addNDesign}
\alias{addNDesign}
\title{Parses probes into nucleotide feature columns to be used in calculating ddG values for binding windows.}
\usage{
addNDesign(data, model, frame)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}

\item{frame}{Position in DNA probe.}
}
\description{
Parses probes into nucleotide feature columns to be used in calculating ddG values for binding windows.
}
