% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{addNewBetas,Intercept-method}
\alias{addNewBetas,Intercept-method}
\title{Add New Betas for \linkS4class{Intercept}}
\usage{
\S4method{addNewBetas}{Intercept}(object, design, value,
  missingValueSuppression, includeDNAstrand, includeView, rcSymmetric)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{design}{Data frame passed to glm that contains the feature design as output by the function addDesignMatrix.}

\item{value}{Glm fit object.}
}
\description{
Updates all slots for object of class \linkS4class{Intercept} using glm model fit.
}
