% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{addNewBetas,model-method}
\alias{addNewBetas,model-method}
\title{Add New Betas to \linkS4class{model}}
\usage{
\S4method{addNewBetas}{model}(object, design, value)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{design}{Design table used to fit glm model.}

\item{value}{Glm fit object.}
}
\value{
\linkS4class{model} object updated to include glm fit data.
}
\description{
Adds new betas to object of class \linkS4class{model} using the output of a glm fit.
}
