% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{addNewBetas,N-method}
\alias{addNewBetas,N-method}
\title{Add New Betas to \linkS4class{N}}
\usage{
\S4method{addNewBetas}{N}(object, design, value, missingValueSuppression,
  useFixedValuesOffset.N)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{design}{Design table used to fit glm model.}

\item{value}{Glm fit object.}

\item{missingValueSuppression}{DdG suppression of features not occurring in the design dataframe relative to the lowest value betas that are fit.}
}
\value{
Object of class \linkS4class{N}, updated to include glm fit data.
}
\description{
Adds new betas to object of class \linkS4class{N} using the output of a glm fit.
}
