% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{addNewBetas,Shape-method}
\alias{addNewBetas,Shape-method}
\title{Add New Betas to \linkS4class{Shape}}
\usage{
\S4method{addNewBetas}{Shape}(object, design, value, useFixedValuesOffset.Shape)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}}

\item{design}{Design table used to fit glm model.}

\item{value}{Glm fit object.}
}
\description{
Adds new betas to \linkS4class{Shape} object using the output of a glm fit.
}
