% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{addSeedPsam<-,model-method}
\alias{addSeedPsam<-,model-method}
\title{Add Seed PSAM}
\usage{
\S4method{addSeedPsam}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{Seeding PSAM for nucleotides.}
}
\description{
Adds seeding nucleotide PSAM to \linkS4class{model} class.
}
