% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{addShapeValues}
\alias{addShapeValues}
\title{Add shape profiles for k-mers.}
\usage{
addShapeValues(data, shapeTable, shape)
}
\arguments{
\item{data}{Data table containing sequences to be profiled. Must include a "Kmer" variable column with character objects.}

\item{shapeTable}{Table with shape parameters to be used for scoring.}

\item{shape}{Shape parameter to be used in scoring.}
}
\description{
Adds Shape parameter values at all offsets of k-mers contained in data table input.
}
