% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{addViewddG}
\alias{addViewddG}
\title{Adds intercept type, view-specific, strand-specific, and round-specific ddG values.}
\usage{
addViewddG(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Adds intercept type, view-specific, strand-specific, and round-specific ddG values.
}
