% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{buildNullEquivalenceMatrix,N-method}
\alias{buildNullEquivalenceMatrix,N-method}
\title{Build Null Equivalence Matrix for \linkS4class{N}}
\usage{
\S4method{buildNullEquivalenceMatrix}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}}
}
\description{
Builds equivalence matrix encoding no symmetries for an object of class \linkS4class{N}.
}
