% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{buildNullEquivalenceMatrix,Shape-method}
\alias{buildNullEquivalenceMatrix,Shape-method}
\title{Build Null Equivalence Matrix for \linkS4class{Shape}}
\usage{
\S4method{buildNullEquivalenceMatrix}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Builds equivalence matrix encoding no symmetries for an object of class \linkS4class{Shape}.
}
