% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRawRegressionFormula.R
\name{buildRawRegressionFormula}
\alias{buildRawRegressionFormula}
\title{Build Raw Regression Formula}
\usage{
buildRawRegressionFormula(model)
}
\arguments{
\item{model}{Object of class \linkS4class{model}.}
}
\value{
Raw regression formula containing features of \linkS4class{model} object.
}
\description{
Builds a raw regression formula for an object of class \linkS4class{model} based on its feature design.
}
\details{
This regression formula must be updated using the data's design matrix to exclude features
in the model that do not appear in the aligned data and/or a single mono-nucleotide at each position
such that the data is not overdescribed.
}
