% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{buildSymmetricEquivalenceMatrix,N-method}
\alias{buildSymmetricEquivalenceMatrix,N-method}
\title{Build Reverse Complement Symmetric Equivalence Matrix for \linkS4class{N}}
\usage{
\S4method{buildSymmetricEquivalenceMatrix}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}}
}
\description{
Builds equivalence matrix encoding reverse complement symmetry for an object of class \linkS4class{N}.
}
