% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{buildSymmetricEquivalenceMatrix,Shape-method}
\alias{buildSymmetricEquivalenceMatrix,Shape-method}
\title{Build Reverse Complement Symmetric Equivalence Matrix for \linkS4class{Shape}}
\usage{
\S4method{buildSymmetricEquivalenceMatrix}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Builds equivalence matrix encoding reverse complement symmetry for an object of class \linkS4class{Shape}.
}
