% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{cleanAffinityInfo}
\alias{cleanAffinityInfo}
\title{Removes view-specific affinity scores from data frame. To be used after selection of probe-specific top binding window.}
\usage{
cleanAffinityInfo(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Removes view-specific affinity scores from data frame. To be used after selection of probe-specific top binding window.
}
