% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{cleanShapeDesign}
\alias{cleanShapeDesign}
\title{Removes shape feature columns used in calculating ddG values for binding windows.}
\usage{
cleanShapeDesign(data, model, frame)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}

\item{frame}{Position in DNA probe.}
}
\description{
Removes shape feature columns used in calculating ddG values for binding windows.
}
