% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaSequenceTools.R
\name{convertSeq2RegEx}
\alias{convertSeq2RegEx}
\title{Convert DNA iupac sequence to regular expression}
\usage{
convertSeq2RegEx(refSeq)
}
\arguments{
\item{refSeq}{Sequence in iupac or regular expression format.}
}
\value{
Sequence in regular expression format recognized by R.
}
\description{
Converts an iupac sequence to R regular expression. If the sequence is already in regular expression form (i.e. it only contains the letters A, C, G, T with brackets and numbers), it will not be changed.
}
