% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaSequenceTools.R
\name{countDNAseqLength}
\alias{countDNAseqLength}
\title{Count sequence length of a DNA regular expression.}
\usage{
countDNAseqLength(reSeq)
}
\arguments{
\item{reSeq}{Sequence in regular expression format (i.e. containing A,C,G,T,'[',']','{','}').}
}
\value{
Length of DNA sequence implied by reSeq input.
}
\description{
Count number of characters implied by a regular expression for DNA sequences.
}
