% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{ddg2affinity}
\alias{ddg2affinity}
\title{Converts view-specific ddG values to affinity scores.}
\usage{
ddg2affinity(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\description{
Converts view-specific ddG values to affinity scores.
}
