% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{formatEquivalenceMatrix,N-method}
\alias{formatEquivalenceMatrix,N-method}
\title{Formats Equivalent Matrix for \linkS4class{N}}
\usage{
\S4method{formatEquivalenceMatrix}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Formats equivalence matrix for object of class \linkS4class{N}, checking for conflicting entries and getting rid of redundancies.
}
