% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{formatEquivalenceMatrix,Shape-method}
\alias{formatEquivalenceMatrix,Shape-method}
\title{Formats Equivalence matrix for \linkS4class{Shape}}
\usage{
\S4method{formatEquivalenceMatrix}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Formats an equivalence matrix for an object of class \linkS4class{Shape}, checking for conflicting instructions and getting rid of redundancies.
}
