% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{g_legend}
\alias{g_legend}
\title{Extract ggplot legend.}
\usage{
g_legend(gplot)
}
\arguments{
\item{gplot}{ggplot object from which to pull legend}
}
\value{
legend ggplot object.
}
\description{
Returns the legend from a ggplot plotting object.
}
