% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getAffinityType,model-method}
\alias{getAffinityType,model-method}
\title{Get Affinity Type}
\usage{
\S4method{getAffinityType}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'affinityType' slot from object of class \linkS4class{model}.
}
