% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getConfidenceLevel,model-method}
\alias{getConfidenceLevel,model-method}
\title{Get Confidence Level}
\usage{
\S4method{getConfidenceLevel}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'confidenceLevel' slot from object of class \linkS4class{model}.
}
