% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getConsensusSeq,model-method}
\alias{getConsensusSeq,model-method}
\title{Get Consensus Sequence}
\usage{
\S4method{getConsensusSeq}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'consensusSeq' slot from object of class \linkS4class{model}.
}
