% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getDateBetasLastAdded,model-method}
\alias{getDateBetasLastAdded,model-method}
\title{Get Date Betas Last Added.}
\usage{
\S4method{getDateBetasLastAdded}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}}
}
\description{
Gets 'dateBetasLastAdded' slot from object of class \linkS4class{model}.
}
