% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getDateInitialized,model-method}
\alias{getDateInitialized,model-method}
\title{Get Date Initialized}
\usage{
\S4method{getDateInitialized}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'dateInitialized' slot from object of class \linkS4class{model}.
}
