% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getDateLastModified,model-method}
\alias{getDateLastModified,model-method}
\title{Get Date Last Modified}
\usage{
\S4method{getDateLastModified}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'dateLastModified' slot from object of class \linkS4class{model}.
}
