% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getDesignMatrix,Intercept-method}
\alias{getDesignMatrix,Intercept-method}
\title{Summary of Design Features for \linkS4class{Intercept}}
\usage{
\S4method{getDesignMatrix}{Intercept}(object, design)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{design}{Data frame with design matrix, as generated by addDesignMatrix function.}
}
\description{
Gets view design from design matrix, i.e. gets counts for each view feature.
}
