% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getDesignMatrix,model-method}
\alias{getDesignMatrix,model-method}
\title{Get design matrix for all features}
\usage{
\S4method{getDesignMatrix}{model}(object, design)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{design}{Data frame with design matrix, as generated by addDesignMatrix function.}
}
\description{
Gets feature design from design matrix.
}
