% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getDesignMatrix,N-method}
\alias{getDesignMatrix,N-method}
\title{Summary Design for Nucleotide Features}
\usage{
\S4method{getDesignMatrix}{N}(object, design)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{design}{Data frame with design matrix, as generated by addDesignMatrix function.}
}
\description{
Gets nucleotide design from design matrix, i.e. gets counts for each mono-nucleotide feature.
}
