% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getDesignMatrix,Shape-method}
\alias{getDesignMatrix,Shape-method}
\title{Design Matrix for \linkS4class{Shape} Features}
\usage{
\S4method{getDesignMatrix}{Shape}(object, design)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{design}{Data frame with design matrix, as generated by \code{addDesignMatrix} function.}
}
\description{
Gets shape design from design matrix, i.e. gets counts for each mono-nucleotide feature.
}
