% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getDesignMatrixSummary,featureSet-method}
\alias{getDesignMatrixSummary,featureSet-method}
\title{Get Design Matrix Summary for \linkS4class{featureSet}}
\usage{
\S4method{getDesignMatrixSummary}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'designMatrixSummary' slot from object of class \linkS4class{featureSet}.
}
