% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getDownFootprintExtend,featureSet-method}
\alias{getDownFootprintExtend,featureSet-method}
\title{Get Length Downstream Footprint Extension}
\usage{
\S4method{getDownFootprintExtend}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'downFootprintExtend' slot from object of class \linkS4class{featureSet}.
}
