% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getDownFootprintExtend,model-method}
\alias{getDownFootprintExtend,model-method}
\title{Get Length Downstream Footprint Extension}
\usage{
\S4method{getDownFootprintExtend}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'downFootprintExtend' slot from object of class \linkS4class{model}.
}
