% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getDownFootprintExtend,N-method}
\alias{getDownFootprintExtend,N-method}
\title{Get Length of Downstream Footprint Extension for Nucleotides}
\usage{
\S4method{getDownFootprintExtend}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.downFootprintExtend slot from object of class \linkS4class{N}.
}
