% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getDownFootprintExtend,Shape-method}
\alias{getDownFootprintExtend,Shape-method}
\title{Get Length (bp) of Downstream Footprint Extension for \linkS4class{Shape}}
\usage{
\S4method{getDownFootprintExtend}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.downFootprintExtend slot from object of class \linkS4class{Shape}.
}
