% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getEquivMat,N-method}
\alias{getEquivMat,N-method}
\title{Get equivalence matrix for \linkS4class{N}}
\usage{
\S4method{getEquivMat}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.equivMat slot from object of class \linkS4class{N}.
}
