% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getEquivMat,Shape-method}
\alias{getEquivMat,Shape-method}
\title{Get Equivalence Matrix for \linkS4class{Shape}}
\usage{
\S4method{getEquivMat}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.equivMat slot from object of class \linkS4class{Shape}.
}
