% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getErrors,Intercept-method}
\alias{getErrors,Intercept-method}
\title{Errors for class \linkS4class{Intercept}}
\usage{
\S4method{getErrors}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.errors slot from object of class \linkS4class{Intercept}.
}
