% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getErrors,N-method}
\alias{getErrors,N-method}
\title{Get beta errors for \linkS4class{N}}
\usage{
\S4method{getErrors}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.errors slot from object of class \linkS4class{N}.
}
