% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getErrors,Shape-method}
\alias{getErrors,Shape-method}
\title{Get Errors for \linkS4class{Shape}}
\usage{
\S4method{getErrors}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.errors slot from object of class \linkS4class{Shape}.
}
