% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getExDownBases,model-method}
\alias{getExDownBases,model-method}
\title{Get Length of Extension Downstream Bases for Shape Parameters}
\usage{
\S4method{getExDownBases}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'exDownBases' slot from object of class \linkS4class{model}
}
