% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getExUpBases,model-method}
\alias{getExUpBases,model-method}
\title{Get Length of Extension Upstream Bases for Shape Parameters}
\usage{
\S4method{getExUpBases}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'exUpBases' slot from object of class \linkS4class{model}
}
