% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getFeatureDesign,featureSet-method}
\alias{getFeatureDesign,featureSet-method}
\title{Get Feature Design for \linkS4class{featureSet}}
\usage{
\S4method{getFeatureDesign}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Defines a print design method for class \linkS4class{featureSet}.
}
