% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getFeatureDesign,Intercept-method}
\alias{getFeatureDesign,Intercept-method}
\title{Get Feature Design for \linkS4class{Intercept}}
\usage{
\S4method{getFeatureDesign}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Defines a feature design print method for class \linkS4class{Intercept}.
}
