% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getFeatureDesign,model-method}
\alias{getFeatureDesign,model-method}
\title{Get Feature Design}
\usage{
\S4method{getFeatureDesign}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Defines a print design method for class \linkS4class{model}.
}
