% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getFeatureDesign,N-method}
\alias{getFeatureDesign,N-method}
\title{Get Feature Design for \linkS4class{N}}
\usage{
\S4method{getFeatureDesign}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Defines a print design method for class \linkS4class{N}.
}
