% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getFeatureDesign,Shape-method}
\alias{getFeatureDesign,Shape-method}
\title{Get Feature Design for \linkS4class{Shape}}
\usage{
\S4method{getFeatureDesign}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Defines a print design method for class \linkS4class{Shape}.
}
