% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getFeatures,model-method}
\alias{getFeatures,model-method}
\title{Get Features}
\usage{
\S4method{getFeatures}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'features' slot from object of class \linkS4class{model}.
}
