% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getFpLen,model-method}
\alias{getFpLen,model-method}
\title{Get Footprint Length}
\usage{
\S4method{getFpLen}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'fpLen' slot from object of class \linkS4class{model}.
}
