% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getFpLen,N-method}
\alias{getFpLen,N-method}
\title{Get Footprint Length}
\usage{
\S4method{getFpLen}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets fpLen slot from object of class \linkS4class{N}.
}
