% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getFpLen,Shape-method}
\alias{getFpLen,Shape-method}
\title{Get Footprint Length for \linkS4class{Shape}}
\usage{
\S4method{getFpLen}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets fpLen slot from object of class \linkS4class{Shape}.
}
