% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getFsDownFootprintExtend,N-method}
\alias{getFsDownFootprintExtend,N-method}
\title{Get Length of Upstream Footprint Extension for Full Feature Set}
\usage{
\S4method{getFsDownFootprintExtend}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets fS.downFootprintExtend slot from object of class \linkS4class{N}.
}
