% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getFsDownFootprintExtend,Shape-method}
\alias{getFsDownFootprintExtend,Shape-method}
\title{Get Length (bp) of Downstream Footprint Extension for Full Feature Set}
\usage{
\S4method{getFsDownFootprintExtend}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets fS.downFootprintExtend slot from object of class \linkS4class{Shape}.
}
