% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getFsUpFootprintExtend,Shape-method}
\alias{getFsUpFootprintExtend,Shape-method}
\title{Get Length (bp) of Upstream Footprint Extension for Full Feature Set}
\usage{
\S4method{getFsUpFootprintExtend}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets fS.upFootprintExtend slot from object of class \linkS4class{Shape}.
}
