% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getGlmFits,featureSet-method}
\alias{getGlmFits,featureSet-method}
\title{Get 'GlmFits' for class \linkS4class{featureSet}}
\usage{
\S4method{getGlmFits}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'GlmFits' slot from object of class \linkS4class{featureSet}.
}
