% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getIncludeDNAstrand,model-method}
\alias{getIncludeDNAstrand,model-method}
\title{Get includeDNAstrand (logical)}
\usage{
\S4method{getIncludeDNAstrand}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'includeDNAstrand' slot from object of class \linkS4class{model}.
}
