% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getIncludeShape,model-method}
\alias{getIncludeShape,model-method}
\title{Get includeShape (logical)}
\usage{
\S4method{getIncludeShape}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'includeShape' slot from object of class \linkS4class{model}.
}
