% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getIncludeView,model-method}
\alias{getIncludeView,model-method}
\title{Get includeView (logical)}
\usage{
\S4method{getIncludeView}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'includeView' slot from object of class \linkS4class{model}.
}
