% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getIntercept,featureSet-method}
\alias{getIntercept,featureSet-method}
\title{Get \linkS4class{Intercept} for \linkS4class{featureSet}}
\usage{
\S4method{getIntercept}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets \linkS4class{Intercept} slot from object of class \linkS4class{featureSet}.
}
