% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getIntercept,model-method}
\alias{getIntercept,model-method}
\title{Get Intercept}
\usage{
\S4method{getIntercept}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'Intercept' slot from object of class \linkS4class{model}.
}
