% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getIteration,model-method}
\alias{getIteration,model-method}
\title{Get Iteration}
\usage{
\S4method{getIteration}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'iteration' slot from object of class \linkS4class{model}.
}
