% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getLeftFixedSeq,model-method}
\alias{getLeftFixedSeq,model-method}
\title{Get Left Fixed Sequence}
\usage{
\S4method{getLeftFixedSeq}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'leftFixedSeq' slot from object of class \linkS4class{model}.
}
