% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getLeftFixedSeqOverlap,model-method}
\alias{getLeftFixedSeqOverlap,model-method}
\title{Get Left Fixed Sequence Overlap}
\usage{
\S4method{getLeftFixedSeqOverlap}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'leftFixedSeqOverlap' slot from object of class \linkS4class{model}.
}
