% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getMinSeedValue,model-method}
\alias{getMinSeedValue,model-method}
\title{Get Minimum Seed Value}
\usage{
\S4method{getMinSeedValue}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'minSeedValue' slot from object of class \linkS4class{model}.
}
