% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getMissingValueSuppression,model-method}
\alias{getMissingValueSuppression,model-method}
\title{Get Missing Value Suppression Factor}
\usage{
\S4method{getMissingValueSuppression}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'missingValueSuppression' slot from object of class \linkS4class{model}.
}
