% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getN,featureSet-method}
\alias{getN,featureSet-method}
\title{Get \linkS4class{N} for \linkS4class{featureSet}}
\usage{
\S4method{getN}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets \linkS4class{N} slot from object of class \linkS4class{featureSet}.
}
