% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getN,model-method}
\alias{getN,model-method}
\title{Get N}
\usage{
\S4method{getN}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'N' slot from object of class \linkS4class{model}.
}
