% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getName,model-method}
\alias{getName,model-method}
\title{Get Name Slot.}
\usage{
\S4method{getName}{model}(object)
}
\arguments{
\item{object}{object of class 'model'}
}
\description{
Gets 'name' slot from object of class \linkS4class{model}.
}
