% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getNumViews,featureSet-method}
\alias{getNumViews,featureSet-method}
\title{Get Number of Views for \linkS4class{featureSet}}
\usage{
\S4method{getNumViews}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'numViews' slot from object of class \linkS4class{featureSet}.
}
