% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getNumViews,model-method}
\alias{getNumViews,model-method}
\title{Get Number of Views per Strand}
\usage{
\S4method{getNumViews}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'numViews' slot from object of class \linkS4class{model}.
}
