% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getOldErrors,Intercept-method}
\alias{getOldErrors,Intercept-method}
\title{Errors for Previous Iterations for \linkS4class{Intercept}}
\usage{
\S4method{getOldErrors}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.oldErrors slot from object of class \linkS4class{Intercept}.
}
