% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getOldErrors,N-method}
\alias{getOldErrors,N-method}
\title{Get beta errors from previous iterations for \linkS4class{N}}
\usage{
\S4method{getOldErrors}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.oldErrors slot from object of class \linkS4class{N}.
}
