% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getOldErrors,Shape-method}
\alias{getOldErrors,Shape-method}
\title{Get Errors from previous iterations for \linkS4class{Shape}}
\usage{
\S4method{getOldErrors}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.oldErrors slot from object of class \linkS4class{Shape}.
}
