% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getOldSig,Intercept-method}
\alias{getOldSig,Intercept-method}
\title{P-values for Previous Iterations for class \linkS4class{Intercept}}
\usage{
\S4method{getOldSig}{Intercept}(object)
}
\arguments{
\item{object}{Object of class '\linkS4class{Intercept}.}
}
\description{
Returns I.oldSig slot from object of class \linkS4class{Intercept}.
}
