% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getOldSig,N-method}
\alias{getOldSig,N-method}
\title{Get beta p-values from previous iterations for \linkS4class{N}}
\usage{
\S4method{getOldSig}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.oldSig slot from object of class \linkS4class{N}.
}
