% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getOldSig,Shape-method}
\alias{getOldSig,Shape-method}
\title{Get P-values from previous iterations for \linkS4class{Shape}}
\usage{
\S4method{getOldSig}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.oldSig slot from object of class \linkS4class{Shape}.
}
