% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getOldValues,Intercept-method}
\alias{getOldValues,Intercept-method}
\title{Values for previous iterations for \linkS4class{Intercept}}
\usage{
\S4method{getOldValues}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.oldValues slot from object of class \linkS4class{Intercept}.
}
