% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getOldValues,N-method}
\alias{getOldValues,N-method}
\title{Get beta values from previous iterations for \linkS4class{N}}
\usage{
\S4method{getOldValues}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.oldValues slot from object of class \linkS4class{N}.
}
