% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getOldValues,Shape-method}
\alias{getOldValues,Shape-method}
\title{Get Values from previous iterations for \linkS4class{Shape}}
\usage{
\S4method{getOldValues}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.oldValues slot from object of class \linkS4class{Shape}.
}
