% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getOldZ,Intercept-method}
\alias{getOldZ,Intercept-method}
\title{Z-scores for Previous Iterations for \linkS4class{Intercept}}
\usage{
\S4method{getOldZ}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.oldZ slot from object of class \linkS4class{Intercept}.
}
