% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getOldZ,N-method}
\alias{getOldZ,N-method}
\title{Get beta z-scores from previous iterations for \linkS4class{N}}
\usage{
\S4method{getOldZ}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.oldZ slot from object of class \linkS4class{N}.
}
