% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getOldZ,Shape-method}
\alias{getOldZ,Shape-method}
\title{Get Z-scores from previous iterations for \linkS4class{Shape}}
\usage{
\S4method{getOldZ}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.oldZ slot from object of class \linkS4class{Shape}.
}
