% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getPlotValues,Intercept-method}
\alias{getPlotValues,Intercept-method}
\title{Gets Plot Values for \linkS4class{Intercept}.}
\usage{
\S4method{getPlotValues}{Intercept}(object, iteration = NULL)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{iteration}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\description{
Reformats I.values and I.errors into a single data.frame for plotting.
}
