% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getPlotValues,N-method}
\alias{getPlotValues,N-method}
\title{Gets plot values for \linkS4class{N}}
\usage{
\S4method{getPlotValues}{N}(object, maxErr = 1, iteration = NULL)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{maxErr}{Maximum y-value to be allowed for error bars in nucleotide dot plot. Truncates error bars that are not contained within this limit.}

\item{iteration}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\value{
Data frame formatted for plotting.
}
\description{
Reformats N.values and N.errors information for plotting.
}
