% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getPlotValues,Shape-method}
\alias{getPlotValues,Shape-method}
\title{Get Plot Values for \linkS4class{Shape}}
\usage{
\S4method{getPlotValues}{Shape}(object, iteration = NULL)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{iteration}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\description{
Reformats Shape.values and Shape.errors into a single data.frame for plotting.
}
