% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getPositions,N-method}
\alias{getPositions,N-method}
\title{Get Positions for Mononucleotides Fitting}
\usage{
\S4method{getPositions}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.set slot from object of class \linkS4class{N}.
}
