% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getPositions,Shape-method}
\alias{getPositions,Shape-method}
\title{Get positions for Shape Parameter Fitting from \linkS4class{Shape}}
\usage{
\S4method{getPositions}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.set slot from object of class \linkS4class{Shape}.
}
