% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbeCounts.R
\name{getProbeCounts}
\alias{getProbeCounts}
\title{Build Probe Count Table}
\usage{
getProbeCounts(SampleA, markovModel, SampleB)
}
\arguments{
\item{SampleA}{SELEX sample handle used to define the probe 'universe'. 
If no \code{SampleB} argument is provided, this sample is first randomly split into two 
equal-sized halves,which are then used to define the counts and the probe universe, 
respectively.}

\item{markovModel}{Markov model for rounds 0.}

\item{SampleB}{SELEX Sample handle used to define the probe counts (optional).}
}
\value{
A data frame containing counts for each probe in the universe.
}
\description{
Create probe count table suitable as starting point for GLM analysis of SELEX data.
}
