% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getPSAM,model-method}
\alias{getPSAM,model-method}
\title{Get Mononucleotide PSAM}
\usage{
\S4method{getPSAM}{model}(object, iteration = NULL)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{iteration}{Iteration of model fit to be used for PSAM}
}
\description{
Gets mononucleotide PSAM (i.e. mono-nucleotide normalized relative affinity contributions) from object of class \linkS4class{model} for specific iteration.
}
\details{
If iteration==NULL, the most recent round of the glm fit values will be used.
}
