% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getPSAM,N-method}
\alias{getPSAM,N-method}
\title{Get Mononucleotide PSAM.}
\usage{
\S4method{getPSAM}{N}(object, iteration = NULL)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{iteration}{Iteration of beta values to be used for PSAM. If iteration == NULL, the most recent round of beta values are used.}
}
\value{
Data.frame containing PSAM of relative affinity values (exp(-ddG))
}
\description{
Reshapes N.values into standard PSAM format.
}
