% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getRcSymmetric,featureSet-method}
\alias{getRcSymmetric,featureSet-method}
\title{Get Reverse Complement Symmetric (logical)}
\usage{
\S4method{getRcSymmetric}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'rcSymmetric' slot from object of class \linkS4class{featureSet}.
}
