% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getRcSymmetric,model-method}
\alias{getRcSymmetric,model-method}
\title{Get Reverse Complement Symmetric (logical)}
\usage{
\S4method{getRcSymmetric}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'rcSymmetric' slot from object of class \linkS4class{model}.
}
