% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{getRegions}
\alias{getRegions}
\title{Get data.frame demarcating regions for a ggplot object.}
\usage{
getRegions(labels, lengths)
}
\arguments{
\item{labels}{labels for each section to be assigned a rectangle}

\item{lengths}{lengths for each of the new sections}
}
\description{
Builds data.frame of regions used to shade different regions of the basic
plots for object of class 'N' and 'Shape'.
}
