% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getRegressionFormula,model-method}
\alias{getRegressionFormula,model-method}
\title{Get Regression Formula}
\usage{
\S4method{getRegressionFormula}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'regressionFormula' slot from object of class \linkS4class{model}.
}
