% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getRightFixedSeq,model-method}
\alias{getRightFixedSeq,model-method}
\title{Get Right Fixed Sequence}
\usage{
\S4method{getRightFixedSeq}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'rightFixedSeq' slot from object of class \linkS4class{model}.
}
