% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getRightFixedSeqOverlap,model-method}
\alias{getRightFixedSeqOverlap,model-method}
\title{Get Right Fixed Sequence Overlap}
\usage{
\S4method{getRightFixedSeqOverlap}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'rightFixedSeqOverlap' slot from object of class \linkS4class{model}.
}
