% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\name{getRoundDesign}
\alias{getRoundDesign}
\title{Summary of rounds design features}
\usage{
getRoundDesign(object, design)
}
\arguments{
\item{object}{object of class 'Intercept'.}

\item{design}{data frame with design matrix, as generated by addDesignMatrix function.}
}
\description{
Gets round design from design matrix, i.e. gets counts for each view feature.
}
