% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getRounds,featureSet-method}
\alias{getRounds,featureSet-method}
\title{Get Rounds}
\usage{
\S4method{getRounds}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'rounds' slot from object of class \linkS4class{featureSet}.
}
