% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getRounds,model-method}
\alias{getRounds,model-method}
\title{Get Rounds}
\usage{
\S4method{getRounds}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'rounds' slot from object of class \linkS4class{model}.
}
