% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getSeedLen,featureSet-method}
\alias{getSeedLen,featureSet-method}
\title{Get Seed Length for \linkS4class{featureSet}}
\usage{
\S4method{getSeedLen}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets seedLen slot from object of class \linkS4class{featureSet}.
}
