% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getSeedLen,model-method}
\alias{getSeedLen,model-method}
\title{Get Seed Length}
\usage{
\S4method{getSeedLen}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'seedLen' slot from object of class \linkS4class{model}.
}
