% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getSeedLen,N-method}
\alias{getSeedLen,N-method}
\title{Get Seed Length}
\usage{
\S4method{getSeedLen}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets seedLen slot from object of class \linkS4class{N}.
}
