% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getSeedLen,Shape-method}
\alias{getSeedLen,Shape-method}
\title{Seed Length for \linkS4class{Shape}}
\usage{
\S4method{getSeedLen}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets seedLen slot from object of class \linkS4class{Shape}.
}
