% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getShape,featureSet-method}
\alias{getShape,featureSet-method}
\title{Get \linkS4class{Shape} for \linkS4class{featureSet}}
\usage{
\S4method{getShape}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets \linkS4class{Shape} slot from object of class \linkS4class{featureSet}.
}
