% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getShape,model-method}
\alias{getShape,model-method}
\title{Get Shape}
\usage{
\S4method{getShape}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'Shape' slot from object of class \linkS4class{model}.
}
