% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getShapeParamsUsed,featureSet-method}
\alias{getShapeParamsUsed,featureSet-method}
\title{Get Shape Parameters Used}
\usage{
\S4method{getShapeParamsUsed}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'shapeParamsUsed' slot from object of class \linkS4class{featureSet}.
}
