% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getShapeParamsUsed,model-method}
\alias{getShapeParamsUsed,model-method}
\title{Get Shape Parameters Used}
\usage{
\S4method{getShapeParamsUsed}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'shapeParamsUsed' slot from object of class \linkS4class{model}.
}
