% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getShapeParamsUsed,Shape-method}
\alias{getShapeParamsUsed,Shape-method}
\title{Get Shape Parameters Used from \linkS4class{Shape}}
\usage{
\S4method{getShapeParamsUsed}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets 'shapeParamsUsed' slot from object of class \linkS4class{Shape}.
}
