% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getShapeTable,model-method}
\alias{getShapeTable,model-method}
\title{Get Shape Table}
\usage{
\S4method{getShapeTable}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'shapeTable' slot from object of class \linkS4class{model}.
}
