% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getSig,Intercept-method}
\alias{getSig,Intercept-method}
\title{P-values for class \linkS4class{Intercept}}
\usage{
\S4method{getSig}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.sig slot from object of class \linkS4class{Intercept}.
}
