% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getSig,N-method}
\alias{getSig,N-method}
\title{Get beta p-values for \linkS4class{N}}
\usage{
\S4method{getSig}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.sig slot from object of class \linkS4class{N}.
}
