% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getSig,Shape-method}
\alias{getSig,Shape-method}
\title{Get P-values for \linkS4class{Shape}}
\usage{
\S4method{getSig}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.sig slot from object of class \linkS4class{Shape}.
}
