% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{getTopRatioLabels}
\alias{getTopRatioLabels}
\title{Get sequence labels for psam ratio plot.}
\usage{
getTopRatioLabels(model1, model2, iter = NULL)
}
\arguments{
\item{model1}{First model used to extract mono-nucleotide feature top ratio labels.}

\item{model2}{Second model used to extract mono-nucleotide feature top ratio labels.}
}
\value{
Vector of nucleotides, with nucleotide at index i being the nucleotide that maximizes the ratio of model1 nucleotide PSAM to model2 nucleotide PSAM at the i'th position.
}
\description{
Gets list of nucleotides maximizing ratio of PSAMs from 2 objects of class 'model' across the models' footprints.
}
