% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getTopSeqLabels,N-method}
\alias{getTopSeqLabels,N-method}
\title{Get Plot Labels for \linkS4class{N} Plot}
\usage{
\S4method{getTopSeqLabels}{N}(object, iteration = NULL)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{iteration}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\value{
List of nucleotides to be used for labeling position in model for nucleotide dot plot.
}
\description{
Gets labels for plotting that correspond to the highest mono-nucleotide affinity sequence.
}
