% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{getUpFootprintExtend,featureSet-method}
\alias{getUpFootprintExtend,featureSet-method}
\title{Get Length Upstream Footprint Extension}
\usage{
\S4method{getUpFootprintExtend}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Gets 'upFootprintExtend' slot from object of class \linkS4class{featureSet}.
}
