% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getUpFootprintExtend,model-method}
\alias{getUpFootprintExtend,model-method}
\title{Get Length Upstream Footprint Extension}
\usage{
\S4method{getUpFootprintExtend}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'upFootprintExtend' slot from object of class \linkS4class{model}.
}
