% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getUpFootprintExtend,N-method}
\alias{getUpFootprintExtend,N-method}
\title{Get Length of Upstream Footprint Extension for Nucleotides}
\usage{
\S4method{getUpFootprintExtend}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.upFootprintExtend slot from object of class \linkS4class{N}.
}
