% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getUpFootprintExtend,Shape-method}
\alias{getUpFootprintExtend,Shape-method}
\title{Get Length (bp) of Upstream Footprint Extension for \linkS4class{Shape}}
\usage{
\S4method{getUpFootprintExtend}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.upFootprintExtend slot from object of class \linkS4class{Shape}.
}
