% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getUseFixedValuesOffset.N,model-method}
\alias{getUseFixedValuesOffset.N,model-method}
\title{Get useFixedValuesOffset.N (logical)}
\usage{
\S4method{getUseFixedValuesOffset.N}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'useFixedValuesOffset.N' slot from object of class \linkS4class{model}.
}
