% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getUseFixedValuesOffset.Shape,model-method}
\alias{getUseFixedValuesOffset.Shape,model-method}
\title{Get useFixedValuesOffset.Shape (logical)}
\usage{
\S4method{getUseFixedValuesOffset.Shape}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'useFixedValuesOffset.Shape' slot from object of class \linkS4class{model}.
}
