% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getValues,Intercept-method}
\alias{getValues,Intercept-method}
\title{Values for \linkS4class{Intercept}}
\usage{
\S4method{getValues}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.values slot from object of class \linkS4class{Intercept}.
}
