% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getValues,N-method}
\alias{getValues,N-method}
\title{Get beta values for \linkS4class{N}}
\usage{
\S4method{getValues}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.values slot from object of class \linkS4class{N}.
}
