% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getValues,Shape-method}
\alias{getValues,Shape-method}
\title{Get Values for \linkS4class{Shape}}
\usage{
\S4method{getValues}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.values slot from object of class \linkS4class{Shape}.
}
