% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getVarRegLen,model-method}
\alias{getVarRegLen,model-method}
\title{Get Variable Region Length}
\usage{
\S4method{getVarRegLen}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'varRegLen' slot from object of class \linkS4class{model}.
}
