% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{getVerbose,model-method}
\alias{getVerbose,model-method}
\title{Get Verbose (logical)}
\usage{
\S4method{getVerbose}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Gets 'verbose' slot from object of class \linkS4class{model}.
}
