% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{getZ,Intercept-method}
\alias{getZ,Intercept-method}
\title{Z-scores for class \linkS4class{Intercept}}
\usage{
\S4method{getZ}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Returns I.z slot from object of class \linkS4class{Intercept}.
}
