% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{getZ,N-method}
\alias{getZ,N-method}
\title{Get beta z-scores for class \linkS4class{N}}
\usage{
\S4method{getZ}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Gets N.z slot from object of class \linkS4class{N}.
}
