% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{getZ,Shape-method}
\alias{getZ,Shape-method}
\title{Get Z-scores for \linkS4class{Shape}}
\usage{
\S4method{getZ}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Gets Shape.z slot from object of class \linkS4class{Shape}.
}
