% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{initialize,featureSet-method}
\alias{initialize,featureSet-method}
\title{Initialize object of class \linkS4class{featureSet}.}
\usage{
\S4method{initialize}{featureSet}(.Object, seedLen, upFootprintExtend,
  downFootprintExtend, rcSymmetric, glmFits, designMatrixSummary, N,
  N.upFootprintExtend, N.downFootprintExtend, N.set, N.values, N.errors, N.z,
  N.sig, N.equivMat, N.oldValues, N.oldErrors, N.oldZ, N.oldSig, Intercept,
  numViews, rounds, I.values, I.errors, I.z, I.sig, I.oldValues, I.oldErrors,
  I.oldZ, I.oldSig, Shape, shapeParamsUsed, Shape.values, Shape.errors, Shape.z,
  Shape.sig, Shape.oldValues, Shape.oldErrors, Shape.oldZ, Shape.oldSig,
  Shape.set, Shape.upFootprintExtend, Shape.downFootprintExtend, Shape.equivMat,
  ...)
}
\arguments{
\item{featureSet}{Object of class \linkS4class{featureSet}.}
}
\description{
Initializes an S4 class to represent full model features.
}
