% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{initialize,Intercept-method}
\alias{initialize,Intercept-method}
\title{Constructor for \linkS4class{Intercept} Class}
\usage{
\S4method{initialize}{Intercept}(.Object, I.values, I.errors, I.z, I.sig,
  I.oldValues, I.oldErrors, I.oldZ, I.oldSig, numViews, rounds, ...)
}
\arguments{
\item{numViews}{number of views scored on each DNA strand.}

\item{rounds}{rounds of SELEX data used to evaluate 'model' to which 'Intercept' belongs.}
}
\description{
Constructor for S4 class  \linkS4class{Intercept}.
}
