% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{initialize,model-method}
\alias{initialize,model-method}
\title{Initialize object of class 'model'}
\usage{
\S4method{initialize}{model}(.Object, name, varRegLen, leftFixedSeq,
  rightFixedSeq, seedLen, consensusSeq, leftFixedSeqOverlap,
  rightFixedSeqOverlap, affinityType, numViews, rounds, confidenceLevel,
  minAffinity, missingValueSuppression, minSeedValue, regressionFormula,
  iteration, upFootprintExtend, downFootprintExtend, rcSymmetric,
  includeDNAstrand, includeView, includeShape, useFixedValuesOffset.N,
  useFixedValuesOffset.Shape, exUpBases, exDownBases, fpLen, shapeParamsUsed,
  shapeTable, features, verbose, dateInitialized, dateBetasLastAdded,
  dateLastModified, N, Intercept, Shape, N.upFootprintExtend,
  N.downFootprintExtend, N.set, N.values, N.errors, N.z, N.sig, N.oldValues,
  N.oldErrors, N.oldZ, N.oldSig, N.equivMat, I.values, I.errors, I.z, I.sig,
  I.oldValues, I.oldErrors, I.oldZ, I.oldSig, Shape.upFootprintExtend,
  Shape.downFootprintExtend, Shape.set, Shape.values, Shape.errors, Shape.z,
  Shape.sig, Shape.oldValues, Shape.oldErrors, Shape.oldZ, Shape.oldSig,
  Shape.equivMat, ...)
}
\arguments{
\item{model}{Object of class 'model'}
}
\description{
Initializes an S4 class to represent full model features
}
