% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{initialize,N-method}
\alias{initialize,N-method}
\title{Constructor for S4 class 'N'}
\usage{
\S4method{initialize}{N}(.Object, seedLen, N.upFootprintExtend,
  N.downFootprintExtend, fS.upFootprintExtend, fS.downFootprintExtend, fpLen,
  N.set, N.equivMat, N.values, N.errors, N.z, N.sig, N.oldValues, N.oldErrors,
  N.oldZ, N.oldSig, rcSymmetric, ...)
}
\description{
Initializes an S4 class to represent mononuclueotide features
}
