% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{initialize,Shape-method}
\alias{initialize,Shape-method}
\title{Constructor for S4 class 'Shape'}
\usage{
\S4method{initialize}{Shape}(.Object, seedLen, Shape.upFootprintExtend,
  Shape.downFootprintExtend, fS.upFootprintExtend, fS.downFootprintExtend,
  fpLen, Shape.set, Shape.equivMat, Shape.values, Shape.errors, Shape.z,
  Shape.sig, Shape.oldValues, Shape.oldErrors, Shape.oldZ, Shape.oldSig,
  shapeParamsUsed, rcSymmetric, ...)
}
\arguments{
\item{shapeParamsUsed}{list of length 1 containing names of shape parameters to be included in model.}

\item{rcSymmetric}{logical: indicates the model described by the 'Shape' object will be reverse complement symmetric.}
}
\description{
Initializes an S4 class to represent shape feature coefficients
}
