% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{class}
\name{Intercept-class}
\alias{Intercept-class}
\alias{Intercept}
\title{'Intercept' Class Definition}
\description{
An S4 class to represent intercept features, including round-, view-, and strand-specific intercept contributions.
}
\section{Slots}{

\describe{
\item{\code{I.values}}{3-D array of beta values with entries corresponding to (Strand)x(View)x(Round) intercept values.}

\item{\code{I.errors}}{3-D array of errors for I.values.}

\item{\code{I.z}}{3-D array of Z-scores for I.values.}

\item{\code{I.sig}}{3-D array of p-values for I.values.}

\item{\code{I.oldValues}}{4-D array for storing previous iteration values beta values for I.values where each entry corresponds to (Strand)x(View)x(Round)x(Iteration).}

\item{\code{I.oldErrors}}{4-D array of previous iteration I.errors.}

\item{\code{I.oldZ}}{4-D array of previous iteration I.z.}

\item{\code{I.oldSig}}{4-D array of previous iteration I.sig.}
}}

