% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateSelexGLM.R
\name{iterateSelexGLM}
\alias{iterateSelexGLM}
\title{Iterates Selex GLM Algorithm}
\usage{
iterateSelexGLM(model, initialData)
}
\arguments{
\item{model}{Object of class \linkS4class{model}.}

\item{initalData}{Table of initial probe counts.}
}
\description{
Loops through SelexGLM algorithm until stable model is reached.
}
