% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{length,featureSet-method}
\alias{length,featureSet-method}
\title{Length of \linkS4class{featureSet}}
\usage{
\S4method{length}{featureSet}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{featureSet}}
}
\description{
Defines 'length' method for object of class \linkS4class{featureSet}.
}
\details{
The length of \linkS4class{featureSet} object is defined as the footprint length of the model.
The footprint length is given by \code{fpLen} = \code{seedLen}+\code{upFootprintExtend}+\code{downFootprintExtend}.
}
