% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{length,Intercept-method}
\alias{length,Intercept-method}
\title{Length for \linkS4class{Intercept}}
\usage{
\S4method{length}{Intercept}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{Intercept}.}
}
\description{
Defines length method for object of class \linkS4class{Intercept}.
}
\details{
Length of \linkS4class{Intercept} object is defined as the number of views scored on each orientation of a DNA strand.
}
