% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{length,model-method}
\alias{length,model-method}
\title{Length of \linkS4class{model}}
\usage{
\S4method{length}{model}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{model}.}
}
\description{
Defines 'length' method for object of class \linkS4class{model}.
}
\details{
The length of \linkS4class{model} object is defined as the footprint length of the model.
The footprint length is given by \code{fpLen} = \code{seedLen}+\code{upFootprintExtend}+\code{downFootprintExtend}.
}
