% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{length,N-method}
\alias{length,N-method}
\title{Length method for \linkS4class{N}}
\usage{
\S4method{length}{N}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{N}.}
}
\value{
Footprint length for mono-nucleotides (i.e. the PSAM length).
}
\description{
Defines a length method for class \linkS4class{N}.
}
