% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{length,Shape-method}
\alias{length,Shape-method}
\title{Length Method for \linkS4class{Shape}}
\usage{
\S4method{length}{Shape}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{Shape}.}
}
\value{
Footprint length for shape parameters.
}
\description{
Defines a length method for class \linkS4class{Shape}.
}
