% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-validity.R
\name{N.validity}
\alias{N.validity}
\title{Validity Function for \linkS4class{N}}
\usage{
N.validity(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Validity Function for object of class \linkS4class{N}.
}
