% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-SelexGLM.R
\docType{package}
\name{NRLBtools}
\alias{NRLBtools}
\alias{NRLBtools-package}
\title{SelexGLM: Utitilies for fitting generalized linear models to SELEX data}
\description{
The SelexGLM package provides tools for inferring protein-DNA recognition models from high-throughout SELEX data using a generalized linear model (GLM) based on Poisson statistics. It was developed by the Bussemaker Lab in the Department of Biological Sciences at Columbia University.
}
\author{
Gabriella D. Martini and Harmen J. Bussemaker
}
