% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\name{plotBasicIntercept}
\alias{plotBasicIntercept}
\title{Plot simple intercept -ddG value for object of class \linkS4class{Intercept}}
\usage{
plotBasicIntercept(x, title = NULL, iter = NULL)
}
\arguments{
\item{x}{Object of class \linkS4class{Intercept}.}

\item{title}{Optional title parameter for plot.}

\item{iter}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\description{
Plots simple intercept -ddG value for object of class \linkS4class{Intercept}.
}
