% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{plotPSAM}
\alias{plotPSAM}
\title{Plot PSAM}
\usage{
plotPSAM(model, title = "", maxErr = 1, iter = NULL, regs = NULL,
  lengths = NULL)
}
\arguments{
\item{model}{model from which to take nucleotide feature values}

\item{title}{title for plot}

\item{maxErr}{maximum height of error bars to use in plotting}

\item{iter}{Iteration of model to use for mono nucleotide values}

\item{regs}{Optional parameter giving labels to use for region shading. If used, lengths parameter must also be specified.}

\item{lengths}{lengths for each of the new sections}
}
\description{
Returns standardized PSAM plot
}
