% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{plotPSAMComp.Ratios}
\alias{plotPSAMComp.Ratios}
\title{Dot plot of ratios between nucleotide psams from 2 objects of class 'model'.}
\usage{
plotPSAMComp.Ratios(model1, model2, title = "", yTitle = "", maxErr = 1,
  iter = NULL, regs = NULL, lengths = NULL)
}
\arguments{
\item{model1}{Model1 from which to take first set of nucleotide feature values.}

\item{model2}{Model2 from which to take first set of nucleotide feature values.}

\item{title}{Title for plot.}

\item{maxErr}{Maximum height of error bars to use in plotting.}

\item{iter}{Iteration of model to use for mono nucleotide values.}

\item{regs}{Optional parameter giving labels to use for region shading. If used, lengths parameter must also be specified.}

\item{lengths}{Length of region parameters to be included.}
}
\description{
Returns standardized PSAM plot of ratios between two model objects' nucleotide affinities.
}
