% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{plotPSAMComp.Scatter}
\alias{plotPSAMComp.Scatter}
\title{Scatter plot showing nucleotide ddG values from 'model1' vs. nucleotide ddG values from 'model2'.}
\usage{
plotPSAMComp.Scatter(model1, model2, title = "", xTitle = "", yTitle = "",
  iter = NULL)
}
\arguments{
\item{model1}{Model1 from which to take first set of nucleotide feature values.}

\item{model2}{Model2 from which to take first set of nucleotide feature values.}

\item{title}{Title for plot.}

\item{xTitle}{Xlabel for plot.}

\item{yTitle}{Ylabel for plot.}

\item{iter}{Iteration of model to use for mono nucleotide values.}

\item{regs}{Optional parameter giving labels to use for region shading. If used, lengths parameter must also be specified.}

\item{lengths}{Length of region parameters to be included.}
}
\description{
Returns standardized scatter plot comparing two sets of mononucleotide features.
}
