% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\name{plotRounds}
\alias{plotRounds}
\title{Plot round-specific -ddG values for object of class \linkS4class{Intercept}}
\usage{
plotRounds(x, sumI, title = NULL, iter = NULL)
}
\arguments{
\item{x}{Object of class \linkS4class{Intercept}.}

\item{title}{Optional title parameter for plot.}

\item{iter}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\description{
Plots round-specific -ddG values for object of class \linkS4class{Intercept}.
}
