% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\name{plotStrands}
\alias{plotStrands}
\title{Plot Strand Offsets for \linkS4class{Intercept}}
\usage{
plotStrands(x, title = NULL, iter = NULL)
}
\arguments{
\item{x}{Object of class \linkS4class{Intercept}.}

\item{title}{Title for plot.}

\item{iter}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\description{
Plots strand-specific affinity contributions.
}
