% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\name{plotViews}
\alias{plotViews}
\title{Plot View-specific ddG contributions for object of class \linkS4class{Intercept}}
\usage{
plotViews(x, rcSymmetric, title = NULL, iter = NULL)
}
\arguments{
\item{x}{Object of class \linkS4class{Intercept}.}

\item{rcSymmetric}{logical: indicates that a reverse complement symmetric model was used for fitting beta values. If missing(rcSymmetric), a value is inferred from the I.values slot of x.}

\item{title}{Optional title parameter for plot.}

\item{iter}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}
}
\description{
Plots view-specific ddG contributions for object of class \linkS4class{Intercept}.
}
\details{
If rcSymmetric is TRUE, view-specific beta values are only modeled for one DNA strand orientation (1x numView - 1 beta values).
If rcSymmetric is FALSE, view-specific beta values are modeled for both DNA strand orientations (2x numView - 1 beta values).
}
