% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{plotZs,Shape-method}
\alias{plotZs,Shape-method}
\title{Plot Z-scores for \linkS4class{Shape}}
\usage{
\S4method{plotZs}{Shape}(object, title = "", iter = NULL, regs = NULL,
  lengths = NULL, verticalPlots = FALSE)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{title}{Title for plot.}

\item{iter}{Iteration of beta values to be used in plot. If iteration == NULL, the most recent round of beta values are used.}

\item{regs}{Optional parameter giving labels to use for region shading. If used, lengths parameter must also be specified.}

\item{lengths}{Length of region parameters to be included.}

\item{verticalPlots}{logical: If TRUE, all plots are stacked vertically. If false, plots are fit into a roughly square block.}

\item{maxErr}{Maximum height of error bars to use in plotting.}
}
\description{
Gets shape plot from object of class \linkS4class{Shape}.
}
