% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{print,featureSet-method}
\alias{print,featureSet-method}
\title{Print method for \linkS4class{featureSet}}
\usage{
\S4method{print}{featureSet}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{featureSet}.}
}
\description{
Defines a print method for \linkS4class{featureSet}.
}
