% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{print,Intercept-method}
\alias{print,Intercept-method}
\title{Print Method for \linkS4class{Intercept}}
\usage{
\S4method{print}{Intercept}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{Intercept}.}
}
\description{
Defines 'print' method for object of class \linkS4class{Intercept}
.
}
