% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{print,model-method}
\alias{print,model-method}
\title{Print \linkS4class{model}}
\usage{
\S4method{print}{model}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{model}.}
}
\description{
Defines a print method for class \linkS4class{model}.
}
