% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{print,N-method}
\alias{print,N-method}
\title{Print method for \linkS4class{N}}
\usage{
\S4method{print}{N}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{N}.}
}
\description{
Defines a print method for class \linkS4class{N}.
}
