% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{print,Shape-method}
\alias{print,Shape-method}
\title{Print Method for \linkS4class{Shape}}
\usage{
\S4method{print}{Shape}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{Shape}.}
}
\description{
Defines a print method for class \linkS4class{Shape}.
}
