% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaSequenceTools.R
\name{revComp}
\alias{revComp}
\title{Gets reverse complement of DNA sequence in iupac or regular expression format.}
\usage{
revComp(refSeq)
}
\arguments{
\item{refSeq}{Sequence in iupac or regular expression format.}
}
\value{
Reverse complement of input in the same format as the input (i.e. regular expression or iupac).
}
\description{
Returns reverse complement of DNA sequence . Extends 'Biostrings' functionality to include regular expression characters and iupac symbols not equal to A, C, G, and T.
}
