% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreKmers.R
\name{scoreKmers}
\alias{scoreKmers}
\title{Score K-mer Sequences}
\usage{
scoreKmers(data, model, l.index, PSAMcol = "PredictedAffinity",
  seqCol = "Kmer")
}
\arguments{
\item{data}{Table of k-mers.}

\item{model}{Object of class \linkS4class{model}.}

\item{l.index}{Left-most position of PSAM to be used for scoring,}

\item{PSAMcol}{Name for PSAM affinity column.}

\item{seqCol}{Name for k-mer variable to be scored.}
}
\description{
Scores Kmers using model beta values for all feature parameters estimated with regression.
}
