% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{seedTable2psam,model-method}
\alias{seedTable2psam,model-method}
\title{Build Seed PSAM}
\usage{
\S4method{seedTable2psam}{model}(object, seedTable)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{seedTable}{Kmer table used to build nucleotide seed.}
}
\value{
Kmer table-derived nucleotide Beta Values.
}
\description{
Uses k-mer table to build PSAM matrix for seeding SelexGLM.
}
