% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{set_panel_size}
\alias{set_panel_size}
\title{Set ggplot panel size}
\usage{
set_panel_size(p = NULL, width = ggplot2::unit(3, "cm"),
  height = ggplot2::unit(3, "cm"))
}
\arguments{
\item{p}{Ggplot object.}

\item{width}{Output width of ggplotGrob object.}

\item{height}{Output height of ggplotGrob object.}
}
\description{
Sets absolute panel size of ggplot objects (not including labels).
}
