% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setAffinityType<-,model-method}
\alias{setAffinityType<-,model-method}
\title{Set Affinity Type}
\usage{
\S4method{setAffinityType}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'affinityType' slot for \linkS4class{model} object.}
}
\description{
Sets 'affinityType' slot from object of class \linkS4class{model}.
}
