% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setConfidenceLevel<-,model-method}
\alias{setConfidenceLevel<-,model-method}
\title{Set Confidence Level}
\usage{
\S4method{setConfidenceLevel}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'confidenceLevel' slot for \linkS4class{model} object.}
}
\description{
Sets 'confidenceLevel' slot from object of class \linkS4class{model}.
}
