% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setConsensusSeq<-,model-method}
\alias{setConsensusSeq<-,model-method}
\title{Set Consensus Sequence}
\usage{
\S4method{setConsensusSeq}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'consensusSeq' slot for \linkS4class{model} object.}
}
\description{
Sets 'consensusSeq' slot for object of class \linkS4class{model}.
}
