% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setDateBetasLastAdded<-,model-method}
\alias{setDateBetasLastAdded<-,model-method}
\title{Set Date Betas Last Added}
\usage{
\S4method{setDateBetasLastAdded}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Sets 'dateBetasLastAdded' slot from object of class \linkS4class{model}.
}
