% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setDateInitialized<-,model-method}
\alias{setDateInitialized<-,model-method}
\title{Set Date Initialized}
\usage{
\S4method{setDateInitialized}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Sets 'dateInitialized' slot from object of class \linkS4class{model}.
}
