% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setDateLastModified<-,model-method}
\alias{setDateLastModified<-,model-method}
\title{Set Date Last Modified}
\usage{
\S4method{setDateLastModified}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Sets 'dateLastModified' slot from object of class \linkS4class{model}.
}
