% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setDesignMatrixSummary<-,featureSet-method}
\alias{setDesignMatrixSummary<-,featureSet-method}
\title{Set Design Matrix Summary for \linkS4class{featureSet}}
\usage{
\S4method{setDesignMatrixSummary}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Sets 'designMatrixSummary' slot from object of class \linkS4class{featureSet}.
}
