% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setDownFootprintExtend<-,featureSet-method}
\alias{setDownFootprintExtend<-,featureSet-method}
\title{Set Length Downstream Footprint Extension}
\usage{
\S4method{setDownFootprintExtend}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{downFootprintExtend for 'model' object to which \linkS4class{featureSet} object belongs.}
}
\description{
Sets 'downFootprintExtend' slot from object of class \linkS4class{featureSet}.
}
