% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setDownFootprintExtend<-,model-method}
\alias{setDownFootprintExtend<-,model-method}
\title{Set Length Downstream Footprint Extension}
\usage{
\S4method{setDownFootprintExtend}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'downFootprintExtend' slot for \linkS4class{model} object.}
}
\description{
Sets 'downFootprintExtend' slot from object of class \linkS4class{model}.
}
