% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setDownFootprintExtend<-,N-method}
\alias{setDownFootprintExtend<-,N-method}
\title{Set Length of Downstream Footprint Extension for Nucleotides}
\usage{
\S4method{setDownFootprintExtend}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Number of bp downstream of seed model to fit mono-nucleotides.}
}
\description{
Sets N.downFootprintExtend slot from object of class \linkS4class{N}.
}
