% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setDownFootprintExtend<-,Shape-method}
\alias{setDownFootprintExtend<-,Shape-method}
\title{Set Length (bp) of Downstream Footprint Extension for \linkS4class{Shape}}
\usage{
\S4method{setDownFootprintExtend}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{number of bp downstream of seed model over which to fit mono-nucleotides for.}
}
\description{
Sets Shape.downFootprintExtend slot for object of class \linkS4class{Shape}.
}
