% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setEquivMat<-,N-method}
\alias{setEquivMat<-,N-method}
\title{Sets Equivalence Matrix for \linkS4class{N}}
\usage{
\S4method{setEquivMat}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Equivalence matrix for object of class \linkS4class{N}.}
}
\description{
Sets an equivalence matrix for object of class \linkS4class{N}.
}
