% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setEquivMat<-,Shape-method}
\alias{setEquivMat<-,Shape-method}
\title{Sets Equivalence Matrix for \linkS4class{Shape}}
\usage{
\S4method{setEquivMat}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Equivalence matrix for object of class \linkS4class{Shape}.}
}
\description{
Sets an equivalence matrix for object of class \linkS4class{Shape}.
}
