% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{setErrors<-,Intercept-method}
\alias{setErrors<-,Intercept-method}
\title{Set Errors for \linkS4class{Intercept}}
\usage{
\S4method{setErrors}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{3-D array with dimension of 2 x (numViews) x (number of rounds) containing -ddG errors for 'Intercept' offsets.}
}
\description{
Sets I.errors slot from object of class \linkS4class{Intercept}.
}
