% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setErrors<-,N-method}
\alias{setErrors<-,N-method}
\title{Set Errors for \linkS4class{N}}
\usage{
\S4method{setErrors}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta errors for mono-nucleotides in -ddG units}
}
\description{
Sets N.errors slot from object of class \linkS4class{N}.
}
