% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setExDownBases<-,model-method}
\alias{setExDownBases<-,model-method}
\title{Set Length of Extension Downstream Bases for Shape Parameters}
\usage{
\S4method{setExDownBases}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'exDownBases' slot for 'model' object}
}
\description{
Sets 'exDownBases' slot for object of class \linkS4class{model}
}
