% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setExUpBases<-,model-method}
\alias{setExUpBases<-,model-method}
\title{Set Length of Extension Upstream Bases for Shape Parameters}
\usage{
\S4method{setExUpBases}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'exUpBases' slot for 'model' object}
}
\description{
Sets 'exUpBases' slot for object of class \linkS4class{model}
}
