% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setFeatures<-,model-method}
\alias{setFeatures<-,model-method}
\title{Set Features}
\usage{
\S4method{setFeatures}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'features' slot for \linkS4class{model} object}
}
\description{
Sets 'features' slot from object of class \linkS4class{model}.
}
