% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setFpLen<-,model-method}
\alias{setFpLen<-,model-method}
\title{Set Footprint Length}
\usage{
\S4method{setFpLen}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'fpLen' slot for \linkS4class{model} object.}
}
\description{
Sets 'fpLen' slot from object of class \linkS4class{model}.
}
