% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setFpLen<-,N-method}
\alias{setFpLen<-,N-method}
\title{Set Footprint Length}
\usage{
\S4method{setFpLen}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Footprint length for full \linkS4class{featureSet} object.}
}
\description{
Sets fpLen slot from object of class \linkS4class{N}.
}
