% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setFpLen<-,Shape-method}
\alias{setFpLen<-,Shape-method}
\title{Set Footprint Length for \linkS4class{Shape}}
\usage{
\S4method{setFpLen}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Footprint length for full \linkS4class{featureSet} object.}
}
\description{
Sets fpLen slot from object of class \linkS4class{Shape}.
}
