% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setFsDownFootprintExtend<-,N-method}
\alias{setFsDownFootprintExtend<-,N-method}
\title{Set Length of Downstream Footprint Extension for Full Feature Set}
\usage{
\S4method{setFsDownFootprintExtend}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Number of bp downstream of seed model to fit any features.}
}
\description{
Sets fS.downFootprintExtend slot from object of class \linkS4class{N}.
}
