% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setFsUpFootprintExtend<-,N-method}
\alias{setFsUpFootprintExtend<-,N-method}
\title{Set Length of Upstream Footprint Extension for Full Feature Set}
\usage{
\S4method{setFsUpFootprintExtend}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Number of bp upstream of seed model to fit any features.}
}
\description{
Sets fS.upFootprintExtend slot from object of class \linkS4class{N}.
}
