% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setFsUpFootprintExtend<-,Shape-method}
\alias{setFsUpFootprintExtend<-,Shape-method}
\title{Set Length (bp) of Upstream Footprint Extension for Full Feature Model}
\usage{
\S4method{setFsUpFootprintExtend}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{number of bp upstream of seed model over which the full 'featureSet' object is defined.}
}
\description{
Sets fS.upFootprintExtend slot for object of class \linkS4class{Shape}.
}
