% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setGlmFits<-,featureSet-method}
\alias{setGlmFits<-,featureSet-method}
\title{Set 'GlmFits' for class \linkS4class{featureSet}}
\usage{
\S4method{setGlmFits}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Sets 'GlmFits' slot from object of class \linkS4class{featureSet}.
}
