% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setIncludeDNAstrand<-,model-method}
\alias{setIncludeDNAstrand<-,model-method}
\title{Set includeDNAstrand (logical)}
\usage{
\S4method{setIncludeDNAstrand}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'includeDNAstrand' slot for \linkS4class{model} object.}
}
\description{
Sets 'includeDNAstrand' slot from object of class \linkS4class{model}.
}
