% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setIncludeShape<-,model-method}
\alias{setIncludeShape<-,model-method}
\title{Set includeShape (logical)}
\usage{
\S4method{setIncludeShape}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'includeShape' slot for \linkS4class{model} object.}
}
\description{
Sets 'includeShape' slot from object of class \linkS4class{model}.
}
