% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setIncludeView<-,model-method}
\alias{setIncludeView<-,model-method}
\title{Set includeView (logical)}
\usage{
\S4method{setIncludeView}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'includeView' slot for \linkS4class{model} object.}
}
\description{
Sets 'includeView' slot from object of class \linkS4class{model}.
}
