% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setIntercept<-,featureSet-method}
\alias{setIntercept<-,featureSet-method}
\title{Set \linkS4class{Intercept} for class \linkS4class{featureSet}}
\usage{
\S4method{setIntercept}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{Object of class \linkS4class{Intercept}.}
}
\description{
Sets \linkS4class{Intercept} slot from object of class \linkS4class{featureSet}.
}
