% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setIntercept<-,model-method}
\alias{setIntercept<-,model-method}
\title{Set Intercept}
\usage{
\S4method{setIntercept}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{Object of class \linkS4class{Intercept}.}
}
\description{
Sets 'Intercept' slot from object of class \linkS4class{model}.
}
