% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setIteration<-,model-method}
\alias{setIteration<-,model-method}
\title{Set Iteration}
\usage{
\S4method{setIteration}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'iteration' slot for \linkS4class{model} object.}
}
\description{
Sets 'iteration' slot from object of class \linkS4class{model}.
}
