% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setLeftFixedSeq<-,model-method}
\alias{setLeftFixedSeq<-,model-method}
\title{Set Left Fixed Sequence}
\usage{
\S4method{setLeftFixedSeq}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'leftFixedSeq' slot for \linkS4class{model} object.}
}
\description{
Sets 'leftFixedSeq' slot for object of class \linkS4class{model}.
}
