% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setLeftFixedSeqOverlap<-,model-method}
\alias{setLeftFixedSeqOverlap<-,model-method}
\title{Set Left Fixed Sequence Overlap}
\usage{
\S4method{setLeftFixedSeqOverlap}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'leftFixedSeqOverlap' slot for \linkS4class{model} object.}
}
\description{
Sets 'leftFixedSeqOverlap' slot from object of class \linkS4class{model}.
}
