% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setMinAffinity<-,model-method}
\alias{setMinAffinity<-,model-method}
\title{Set Minimum Affinity}
\usage{
\S4method{setMinAffinity}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'minAffinity' slot for \linkS4class{model} object.}
}
\description{
Sets 'minAffinity' slot from object of class \linkS4class{model}.
}
