% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setMinSeedValue<-,model-method}
\alias{setMinSeedValue<-,model-method}
\title{Set Minimum Seed Value}
\usage{
\S4method{setMinSeedValue}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'minSeedValue' slot for \linkS4class{model} object.}
}
\description{
Sets 'minSeedValue' slot from object of class \linkS4class{model}.
}
