% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setMissingValueSuppression<-,model-method}
\alias{setMissingValueSuppression<-,model-method}
\title{Set Missing Value Suppression Factor}
\usage{
\S4method{setMissingValueSuppression}{model}(object) <- value
}
\arguments{
\item{object}{object of class \linkS4class{model}.}

\item{value}{'missingValueSuppression' slot for \linkS4class{model} object.}
}
\description{
Sets 'missingValueSuppression' slot from object of class \linkS4class{model}.
}
