% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setN<-,featureSet-method}
\alias{setN<-,featureSet-method}
\title{Set \linkS4class{N} for class \linkS4class{featureSet}}
\usage{
\S4method{setN}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{Object of class \linkS4class{N}.}
}
\description{
Sets \linkS4class{N} slot from object of class \linkS4class{featureSet}.
}
