% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setN<-,model-method}
\alias{setN<-,model-method}
\title{Set N}
\usage{
\S4method{setN}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{Object of class \linkS4class{N}.}
}
\description{
Sets 'N' slot from object of class \linkS4class{model}.
}
