% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setName<-,model-method}
\alias{setName<-,model-method}
\title{Sets name slot.}
\usage{
\S4method{setName}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'name' slot from \linkS4class{model}.}
}
\description{
Sets 'name' slot from object of class \linkS4class{model}.
}
