% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setNumViews<-,featureSet-method}
\alias{setNumViews<-,featureSet-method}
\title{Set Number of Views for \linkS4class{featureSet}}
\usage{
\S4method{setNumViews}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{NumViews for \linkS4class{model} object to which \linkS4class{featureSet} object belongs}
}
\description{
Sets numViews slot from object of class \linkS4class{featureSet}.
}
