% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setNumViews<-,model-method}
\alias{setNumViews<-,model-method}
\title{Set Number of Views per Strand}
\usage{
\S4method{setNumViews}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'numViews' slot for \linkS4class{model} object}
}
\description{
Sets 'numViews' slot from object of class \linkS4class{model}.
}
