% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{setOldErrors<-,Intercept-method}
\alias{setOldErrors<-,Intercept-method}
\title{Set Errors for Previous Iterations for \linkS4class{Intercept}}
\usage{
\S4method{setOldErrors}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{4-D array with dimension of 2 x (numViews) x (number of rounds) x (number of glm iterations) containing -ddG errors for 'Intercept' offsets.}
}
\description{
Sets I.oldErrors slot for object of class \linkS4class{Intercept}.
}
