% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setOldSig<-,N-method}
\alias{setOldSig<-,N-method}
\title{Set P-values for Previous Iterations for \linkS4class{N}}
\usage{
\S4method{setOldSig}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Array of 4 x fpLen x (iterations-1) containing beta p-values for mono-nucleotides in -ddG units}
}
\description{
Sets N.oldSig slot from object of class \linkS4class{N}.
}
