% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setOldSig<-,Shape-method}
\alias{setOldSig<-,Shape-method}
\title{Set P-values from Previous Iterations for \linkS4class{Shape}}
\usage{
\S4method{setOldSig}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Array of length(shapeParamsUsed[[1]]) x fpLen x (iterations-1) containing beta p-values for shape coefficients in -ddG/angstrom units.}
}
\description{
Sets Shape.oldSig slot from object of class \linkS4class{Shape}.
}
