% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setOldValues<-,N-method}
\alias{setOldValues<-,N-method}
\title{Set Values for Previous Iterations for \linkS4class{N}}
\usage{
\S4method{setOldValues}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Array of 4 x fpLen x (iterations-1) containing beta values for mono-nucleotides in -ddG units}
}
\description{
Sets N.oldValues slot from object of class \linkS4class{N}.
}
