% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setOldValues<-,Shape-method}
\alias{setOldValues<-,Shape-method}
\title{Set Values for Previous Iterations for \linkS4class{Shape}}
\usage{
\S4method{setOldValues}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Array of length(shapeParamsUsed[[1]]) x fpLen x (iterations-1) containing beta values for shape coefficients in -ddG/angstrom units.}
}
\description{
Sets Shape.oldValues slot from object of class \linkS4class{Shape}.
}
