% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setOldZ<-,N-method}
\alias{setOldZ<-,N-method}
\title{Set Z-scores for Previous Iterations for \linkS4class{N}}
\usage{
\S4method{setOldZ}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Array of 4 x fpLen x (iterations-1) containing beta z-scores for mono-nucleotides in -ddG units}
}
\description{
Sets N.oldZ slot from object of class \linkS4class{N}.
}
