% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setOldZ<-,Shape-method}
\alias{setOldZ<-,Shape-method}
\title{Set Z-scores for Previous Iterations for \linkS4class{Shape}}
\usage{
\S4method{setOldZ}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Array of length(shapeParamsUsed[[1]]) x fpLen x (iterations-1) containing beta z-scores for shape coefficients in -ddG/angstrom units.}
}
\description{
Sets Shape.oldZ slot from object of class \linkS4class{Shape}.
}
