% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setPositions<-,N-method}
\alias{setPositions<-,N-method}
\title{Set Positions for Mononucleotides for \linkS4class{N}}
\usage{
\S4method{setPositions}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Vector of positions within \linkS4class{featureSet} for which mononucleotide features should be fit.}
}
\description{
Sets N.set slot from object of class \linkS4class{N}.
}
