% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setPositions<-,Shape-method}
\alias{setPositions<-,Shape-method}
\title{Set positions for Shape Parameter Fitting from \linkS4class{Shape}}
\usage{
\S4method{setPositions}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Vector of positions within \linkS4class{featureSet} for which shape parameter coefficients should be fit.}
}
\description{
Sets Shape.set slot for object of class \linkS4class{Shape}.
}
