% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setRcSymmetric<-,featureSet-method}
\alias{setRcSymmetric<-,featureSet-method}
\title{Set Reverse Complement Symmetry (logical)}
\usage{
\S4method{setRcSymmetric}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{RcSymmetric for \linkS4class{model} object to which \linkS4class{featureSet} object belongs.}
}
\description{
Sets rcSymmetric slot from object of class \linkS4class{featureSet}.
}
