% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setRcSymmetric<-,model-method}
\alias{setRcSymmetric<-,model-method}
\title{Set Reverse Complement Symmetric (logical)}
\usage{
\S4method{setRcSymmetric}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'rcSymmetric' slot for \linkS4class{model} object.}
}
\description{
Sets 'rcSymmetric' slot from object of class \linkS4class{model}.
}
