% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setRegressionFormula<-,model-method}
\alias{setRegressionFormula<-,model-method}
\title{Set Regression Formula}
\usage{
\S4method{setRegressionFormula}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'regressionFormula' slot for \linkS4class{model} object.}
}
\description{
Sets 'regressionFormula' slot from object of class \linkS4class{model}.
}
