% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setRightFixedSeq<-,model-method}
\alias{setRightFixedSeq<-,model-method}
\title{Set Right Fixed Sequence}
\usage{
\S4method{setRightFixedSeq}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'rightFixedSeq' slot for \linkS4class{model} object.}
}
\description{
Sets 'rightFixedSeq' slot from object of class \linkS4class{model}.
}
