% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setRightFixedSeqOverlap<-,model-method}
\alias{setRightFixedSeqOverlap<-,model-method}
\title{Set Right Fixed Sequence Overlap}
\usage{
\S4method{setRightFixedSeqOverlap}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'rightFixedSeqOverlap' slot for \linkS4class{model} object.}
}
\description{
Sets 'rightFixedSeqOverlap' slot from object of class \linkS4class{model}.
}
