% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setRounds<-,featureSet-method}
\alias{setRounds<-,featureSet-method}
\title{Set Rounds for \linkS4class{featureSet}}
\usage{
\S4method{setRounds}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{Rounds for \linkS4class{model} object to which \linkS4class{featureSet} object belongs.}
}
\description{
Sets rounds slot from object of class \linkS4class{featureSet}.
}
