% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setRounds<-,model-method}
\alias{setRounds<-,model-method}
\title{Set Rounds}
\usage{
\S4method{setRounds}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}}

\item{value}{'rounds' slot for \linkS4class{model} object.}
}
\description{
Sets 'rounds' slot from object of class \linkS4class{model}.
}
