% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setSeedLen<-,featureSet-method}
\alias{setSeedLen<-,featureSet-method}
\title{Set Seed Length for \linkS4class{featureSet}}
\usage{
\S4method{setSeedLen}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{Seed length for 'model' object to which \linkS4class{featureSet} object belongs.}
}
\description{
Sets seedLen slot from object of class \linkS4class{featureSet}.
}
