% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setSeedLen<-,model-method}
\alias{setSeedLen<-,model-method}
\title{Set Seed Length}
\usage{
\S4method{setSeedLen}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'seedLen' slot for \linkS4class{model} object.}
}
\description{
Sets 'seedLen' slot from object of class \linkS4class{model}.
}
