% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setSeedLen<-,N-method}
\alias{setSeedLen<-,N-method}
\title{Set Seed Length}
\usage{
\S4method{setSeedLen}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Seed length for 'model' object to which \linkS4class{N} object belongs.}
}
\description{
Sets seedLen slot from object of class \linkS4class{N}.
}
