% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setSeedLen<-,Shape-method}
\alias{setSeedLen<-,Shape-method}
\title{Set Seed Length for \linkS4class{Shape}}
\usage{
\S4method{setSeedLen}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Seed length for \linkS4class{model} object to which \linkS4class{Shape} object belongs.}
}
\description{
Sets seedLen slot from object of class \linkS4class{Shape}.
}
