% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setShape<-,featureSet-method}
\alias{setShape<-,featureSet-method}
\title{Set \linkS4class{Shape} for class \linkS4class{featureSet}}
\usage{
\S4method{setShape}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{Object of class \linkS4class{Shape}.}
}
\description{
Sets \linkS4class{Shape} slot from object of class \linkS4class{featureSet}.
}
