% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setShape<-,model-method}
\alias{setShape<-,model-method}
\title{Set Shape}
\usage{
\S4method{setShape}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{Object of class \linkS4class{Shape}.}
}
\description{
Sets 'Shape' slot from object of class \linkS4class{model}.
}
