% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setShapeParamsUsed<-,featureSet-method}
\alias{setShapeParamsUsed<-,featureSet-method}
\title{Set Shape Parameters Used for \linkS4class{featureSet}}
\usage{
\S4method{setShapeParamsUsed}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{ShapeParamsUsed for \linkS4class{model} object to which \linkS4class{featureSet} object belongs.}
}
\description{
Sets shapeParamsUsed slot from object of class \linkS4class{featureSet}.
}
