% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setShapeParamsUsed<-,model-method}
\alias{setShapeParamsUsed<-,model-method}
\title{Set Shape Parameters Used}
\usage{
\S4method{setShapeParamsUsed}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'shapeParamsUsed' slot for \linkS4class{model} object.}
}
\description{
Sets 'shapeParamsUsed' slot from object of class \linkS4class{model}.
}
