% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setShapeParamsUsed<-,Shape-method}
\alias{setShapeParamsUsed<-,Shape-method}
\title{Set Shape Parameters Used for \linkS4class{Shape}}
\usage{
\S4method{setShapeParamsUsed}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{ShapeParamsUsed for \linkS4class{model} object to which \linkS4class{Shape} object belongs.}
}
\description{
Sets shapeParamsUsed slot from object of class \linkS4class{Shape}.
}
