% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setShapeTable<-,model-method}
\alias{setShapeTable<-,model-method}
\title{Set Shape Table}
\usage{
\S4method{setShapeTable}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'shapeTable' slot for \linkS4class{model} object.}
}
\description{
Sets 'shapeTable' slot from object of class \linkS4class{model}.
}
