% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{setSig<-,Intercept-method}
\alias{setSig<-,Intercept-method}
\title{Set P-values for \linkS4class{Intercept}}
\usage{
\S4method{setSig}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{3-D array with dimension of 2 x (numViews) x (number of rounds) containing -ddG p-values for 'Intercept' offsets.}
}
\description{
Sets I.sig slot from object of class \linkS4class{Intercept}.
}
