% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setSig<-,N-method}
\alias{setSig<-,N-method}
\title{Set P-values for \linkS4class{N}}
\usage{
\S4method{setSig}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta p-values for mono-nucleotides in -ddG units}
}
\description{
Sets N.sig slot from object of class \linkS4class{N}.
}
