% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setSig<-,Shape-method}
\alias{setSig<-,Shape-method}
\title{Set P-values for \linkS4class{Shape}}
\usage{
\S4method{setSig}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of length(shapeParamsUsed[[1]]) x fpLen containing beta sig for shape coefficients in -ddG/angstrom units.}
}
\description{
Sets Shape.sig slot from object of class \linkS4class{Shape}.
}
