% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{setUpFootprintExtend<-,featureSet-method}
\alias{setUpFootprintExtend<-,featureSet-method}
\title{Set Length Upstream Footprint Extension}
\usage{
\S4method{setUpFootprintExtend}{featureSet}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}

\item{value}{upFootprintExtend for 'model' object to which \linkS4class{featureSet} object belongs.}
}
\description{
Sets 'upFootprintExtend' slot from object of class \linkS4class{featureSet}.
}
