% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setUpFootprintExtend<-,model-method}
\alias{setUpFootprintExtend<-,model-method}
\title{Set Length Upstream Footprint Extension}
\usage{
\S4method{setUpFootprintExtend}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'upFootprintExtend' slot for \linkS4class{model} object.}
}
\description{
Sets 'upFootprintExtend' slot from object of class \linkS4class{model}.
}
