% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setUpFootprintExtend<-,N-method}
\alias{setUpFootprintExtend<-,N-method}
\title{Set Length of Upstream Footprint Extension for Nucleotides}
\usage{
\S4method{setUpFootprintExtend}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Number of bp upstream of seed model to fit mono-nucleotides.}
}
\description{
Sets N.upFootprintExtend slot from object of class \linkS4class{N}.
}
