% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setUpFootprintExtend<-,Shape-method}
\alias{setUpFootprintExtend<-,Shape-method}
\title{Set Length (bp) of Upstream Footprint Extension for \linkS4class{Shape}}
\usage{
\S4method{setUpFootprintExtend}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Number of bp upstream of seed model to fit shape coefficients for.}
}
\description{
Sets Shape.upFootprintExtend slot for object of class \linkS4class{Shape}.
}
