% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setUseFixedValuesOffset.N<-,model-method}
\alias{setUseFixedValuesOffset.N<-,model-method}
\title{Set 'useFixedValuesOffset.N' (logical)}
\usage{
\S4method{setUseFixedValuesOffset.N}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'useFixedValuesOffset.N'' slot for \linkS4class{model} object.}
}
\description{
Sets 'useFixedValuesOffset.N' slot from object of class \linkS4class{model}.
}
