% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setUseFixedValuesOffset.Shape<-,model-method}
\alias{setUseFixedValuesOffset.Shape<-,model-method}
\title{Set useFixedValuesOffset.Shape (logical)}
\usage{
\S4method{setUseFixedValuesOffset.Shape}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'useFixedValuesOffset.Shape'' slot for \linkS4class{model} object}
}
\description{
Sets 'useFixedValuesOffset.Shape' slot from object of class \linkS4class{model}.
}
