% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{setValues<-,Intercept-method}
\alias{setValues<-,Intercept-method}
\title{Set Values for class \linkS4class{Intercept}}
\usage{
\S4method{setValues}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{3-D array with dimension of 2 x (numViews) x (number of rounds) containing -ddG values for 'Intercept' offsets.}
}
\description{
Sets I.values slot from object of class \linkS4class{Intercept}.
}
