% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setValues<-,Shape-method}
\alias{setValues<-,Shape-method}
\title{Set Values for \linkS4class{Shape}}
\usage{
\S4method{setValues}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of length(shapeParamsUsed[[1]]) x fpLen containing beta values for shape coefficients in -ddG/angstrom units.}
}
\description{
Sets Shape.values slot from object of class \linkS4class{Shape}.
}
