% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setVarRegLen<-,model-method}
\alias{setVarRegLen<-,model-method}
\title{Set Variable Region Length.}
\usage{
\S4method{setVarRegLen}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}}

\item{value}{'varRegLen' slot for \linkS4class{model} object.}
}
\description{
Sets 'varRegLen' slot from object of class \linkS4class{model}.
}
