% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{setVerbose<-,model-method}
\alias{setVerbose<-,model-method}
\title{Set verbose (logical)}
\usage{
\S4method{setVerbose}{model}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}

\item{value}{'verbose' slot for \linkS4class{model} object.}
}
\description{
Sets 'verbose' slot from object of class \linkS4class{model}.
}
