% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{setZ<-,Intercept-method}
\alias{setZ<-,Intercept-method}
\title{Set Z-scores for \linkS4class{Intercept}}
\usage{
\S4method{setZ}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{3-D array with dimension of 2 x (numViews) x (number of rounds) containing -ddG z-scores for 'Intercept' offsets.}
}
\description{
Sets I.z slot from object of class \linkS4class{Intercept}.
}
