% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{setZ<-,N-method}
\alias{setZ<-,N-method}
\title{Set Z-scores for \linkS4class{N}}
\usage{
\S4method{setZ}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta z-scores for mono-nucleotides in -ddG units}
}
\description{
Sets N.z slot from object of class \linkS4class{N}.
}
