% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{setZ<-,Shape-method}
\alias{setZ<-,Shape-method}
\title{Set Z-scores for \linkS4class{Shape}}
\usage{
\S4method{setZ}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of length(shapeParamsUsed[[1]]) x fpLen containing beta z for shape coefficients in -ddG/angstrom units.}
}
\description{
Sets Shape.z slot from object of class \linkS4class{Shape}.
}
