% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\name{shapePlotRows}
\alias{shapePlotRows}
\title{Get Number of Rows for \linkS4class{Shape} Plot}
\usage{
shapePlotRows(x, verticalPlots = FALSE)
}
\arguments{
\item{x}{Object of class \linkS4class{Shape}.}

\item{verticalPlots}{logical: indicates that only one shape parameter should be plotted per row. This format is ideal for markdown, where the height of the output plot can be very flexibly accommodated.}
}
\description{
Compute number of rows to be used with facet_wrap for plotting beta values for objects of class 'Shape'.
Number of rows vary depending on the number of shape parameters fit for the object of class 'Shape'.
}
