% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\name{shapeXlabSpacing}
\alias{shapeXlabSpacing}
\title{Get Spacing for X-labels for \linkS4class{Shape} Plot}
\usage{
shapeXlabSpacing(x, sPs)
}
\arguments{
\item{x}{Object of class \linkS4class{Shape}.}

\item{sPs}{Number of rows to be used for plot(\linkS4class{Shape}) (output of \code{shapePlotRows}).}
}
\description{
Gets spacing for x-labels used for shape plotting based on the number of shape parameters per row.
}
