% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{show,featureSet-method}
\alias{show,featureSet-method}
\title{Show method for class 'featureSet'}
\usage{
\S4method{show}{featureSet}(object)
}
\arguments{
\item{object}{object of class 'featureSet'.}
}
\description{
Defines a show method for class 'featureSet'.
}
