% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{show,Intercept-method}
\alias{show,Intercept-method}
\title{Show Method for \linkS4class{Intercept}}
\usage{
\S4method{show}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Defines show method for \linkS4class{Intercept} class.
}
