% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{show,model-method}
\alias{show,model-method}
\title{Show \linkS4class{model}}
\usage{
\S4method{show}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Defines a show method for class \linkS4class{model}.
}
