% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{show,N-method}
\alias{show,N-method}
\title{Show method for \linkS4class{N}}
\usage{
\S4method{show}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Defines a show method for class \linkS4class{N}.
}
