% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{show,Shape-method}
\alias{show,Shape-method}
\title{Show method for \linkS4class{Shape}}
\usage{
\S4method{show}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Defines a show method for class \linkS4class{Shape}.
}
