% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-class.R
\docType{methods}
\name{summary,featureSet-method}
\alias{summary,featureSet-method}
\title{Summary Method for \linkS4class{featureSet}}
\usage{
\S4method{summary}{featureSet}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Defines a summary method for class \linkS4class{featureSet}.
}
