% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{summary,Intercept-method}
\alias{summary,Intercept-method}
\title{Summary Method for  \linkS4class{Intercept}}
\usage{
\S4method{summary}{Intercept}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Defines summary method for object of class  \linkS4class{Intercept}.
}
