% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{summary,model-method}
\alias{summary,model-method}
\title{Summary method for \linkS4class{model}}
\usage{
\S4method{summary}{model}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Defines a summary method for class \linkS4class{model}.
}
