% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{summary,N-method}
\alias{summary,N-method}
\title{Summary method for \linkS4class{N}}
\usage{
\S4method{summary}{N}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}
}
\description{
Defines a summary method for class \linkS4class{N}.
}
