% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{summary,Shape-method}
\alias{summary,Shape-method}
\title{Summary method for \linkS4class{Shape}}
\usage{
\S4method{summary}{Shape}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Defines a summary method for class \linkS4class{Shape}.
}
