% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-validity.R
\name{testNBetaSymmetry}
\alias{testNBetaSymmetry}
\title{Test Symmetry of Nucleotide Betas}
\usage{
testNBetaSymmetry(mat, NSet)
}
\arguments{
\item{mat}{Matrix of nucleotide beta values from \linkS4class{N}.}

\item{Nset}{Set of nucleotide values included in model (i.e. N.set from \linkS4class{N}).}
}
\description{
Tests reverse complement symmetry of nucleotide betas for nucleotides in N.set
}
