% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-validity.R
\name{testShapeBetaSymmetry}
\alias{testShapeBetaSymmetry}
\title{Test Symmetry of Shape Betas}
\usage{
testShapeBetaSymmetry(mat, SSet)
}
\arguments{
\item{mat}{Matrix of shape beta values from \linkS4class{Shape}.}

\item{Sset}{Set of shape values included in model (i.e. Shape.set from \linkS4class{Shape}).}
}
\description{
Tests reverse complement symmetry of shape betas for nucleotides in Shape.set
}
