% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-validity.R
\name{testViewBetaSymmetry}
\alias{testViewBetaSymmetry}
\title{Test Symmetry of View Betas}
\usage{
testViewBetaSymmetry(arr)
}
\arguments{
\item{arr}{Array of beta values from \linkS4class{Intercept}.}
}
\description{
Tests reverse complement symmetry of view betas.
}
