% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topModelMatch.R
\name{topModelMatch}
\alias{topModelMatch}
\title{Finds Top Match View on Probe for Model.}
\usage{
topModelMatch(data, model)
}
\arguments{
\item{data}{Count table for DNA probes used for filtering.}

\item{model}{Object of class \linkS4class{model}.}
}
\value{
A data frame containing probes with top binding windows selected, along with their PSAM scores and confidence levels.
}
\description{
Selects top binding window for each probe and assigns model affinity score and confidence level for selected window.
}
