% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatedRegressionFormula.R
\name{updatedRegressionFormula}
\alias{updatedRegressionFormula}
\title{Get Updated Regression Formula}
\usage{
updatedRegressionFormula(design, model)
}
\arguments{
\item{design}{Design Matrix for regression.}

\item{model}{Object of class \linkS4class{model} used to score binding windows and direct analysis.}
}
\value{
Regression formula of class character.
}
\description{
Uses design matrix in \code{design} to build regression formula with independent nucleotide features.
Additionally, removes features with the same value for every observation in \code{design}.
}
