% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{updateErrors<-,Intercept-method}
\alias{updateErrors<-,Intercept-method}
\title{Update Errors for \linkS4class{Intercept}}
\usage{
\S4method{updateErrors}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{Glm fit object.}
}
\description{
Updates I.errors and I.oldErrors slots for object of class \linkS4class{Intercept} using output of glm model fit.
}
