% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{updateErrors<-,N-method}
\alias{updateErrors<-,N-method}
\title{Update Errors for \linkS4class{N}}
\usage{
\S4method{updateErrors}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta errors for mono-nucleotides in -ddG units.}
}
\description{
Updates N.errors slot from object of class \linkS4class{N}.
}
