% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{updateErrors<-,Shape-method}
\alias{updateErrors<-,Shape-method}
\title{Update Errors for \linkS4class{Shape}}
\usage{
\S4method{updateErrors}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of 4xfpLen containing beta errors for shape coefficients in -ddG/angstrom units.}
}
\description{
Updates Shape.errors slot from object of class \linkS4class{Shape}.
}
