% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{updateSig<-,Intercept-method}
\alias{updateSig<-,Intercept-method}
\title{Update P-values for \linkS4class{Intercept}}
\usage{
\S4method{updateSig}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{Glm fit object.}
}
\description{
Updates I.sig and I.oldSig slots for object of class \linkS4class{Intercept} using output of glm model fit.
}
