% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{updateSig<-,N-method}
\alias{updateSig<-,N-method}
\title{Update P-values for class \linkS4class{N}}
\usage{
\S4method{updateSig}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta p-values for mono-nucleotides in -ddG units.}
}
\description{
Updates N.sig slot from object of class \linkS4class{N}.
}
