% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{updateSig<-,Shape-method}
\alias{updateSig<-,Shape-method}
\title{Update P-values for class \linkS4class{Shape}}
\usage{
\S4method{updateSig}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of 4xfpLen containing beta p-values for shape coefficients in -ddG/angstrom units.}
}
\description{
Updates Shape.sig slot from object of class \linkS4class{Shape}
}
