% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{updateValues<-,Intercept-method}
\alias{updateValues<-,Intercept-method}
\title{Update Values for \linkS4class{Intercept}}
\usage{
\S4method{updateValues}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{Glm fit object.}
}
\description{
Updates I.values and I.oldValues slots for object of class \linkS4class{Intercept} using output of glm model fit.
}
