% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{updateValues<-,N-method}
\alias{updateValues<-,N-method}
\title{Update Values for \linkS4class{N}}
\usage{
\S4method{updateValues}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta values for mono-nucleotides in -ddG units.}
}
\description{
Updates N.values slot from object of class \linkS4class{N}.
}
