% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{updateValues<-,Shape-method}
\alias{updateValues<-,Shape-method}
\title{Update Values for \linkS4class{Shape}}
\usage{
\S4method{updateValues}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of 4xfpLen containing beta values for shape coefficients in -ddG/angstrom units.}
}
\description{
Updates Shape.values slot from object of class \linkS4class{Shape}.
}
