% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-class.R
\docType{methods}
\name{updateZ<-,Intercept-method}
\alias{updateZ<-,Intercept-method}
\title{Update Z-scores for \linkS4class{Intercept}}
\usage{
\S4method{updateZ}{Intercept}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}

\item{value}{Glm fit object.}
}
\description{
Updates I.z and I.oldZ slots for object of class \linkS4class{Intercept} using output of glm model fit.
}
