% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-class.R
\docType{methods}
\name{updateZ<-,N-method}
\alias{updateZ<-,N-method}
\title{Update Z-scores for \linkS4class{N}}
\usage{
\S4method{updateZ}{N}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{N}.}

\item{value}{Matrix of 4xfpLen containing beta z-scores for mono-nucleotides in -ddG units.}
}
\description{
Updates N.z slot from object of class \linkS4class{N}.
}
