% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-class.R
\docType{methods}
\name{updateZ<-,Shape-method}
\alias{updateZ<-,Shape-method}
\title{Update Z-scores for \linkS4class{Shape}}
\usage{
\S4method{updateZ}{Shape}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}

\item{value}{Matrix of 4xfpLen containing beta z-scores for shape coefficients in -ddG/angstrom units.}
}
\description{
Updates Shape.z slot from object of class \linkS4class{Shape}.
}
