% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSet-validity.R
\name{validFeatureSet}
\alias{validFeatureSet}
\title{Validity Function for \linkS4class{featureSet}}
\usage{
validFeatureSet(object)
}
\arguments{
\item{object}{Object of class \linkS4class{featureSet}.}
}
\description{
Validity Function for object of class \linkS4class{featureSet}.
}
