% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intercept-validity.R
\name{validIntercept}
\alias{validIntercept}
\title{Validity Function for \linkS4class{Intercept}}
\usage{
validIntercept(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Intercept}.}
}
\description{
Validity Function for object of class \linkS4class{Intercept}.
}
