% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-validity.R
\name{validModel}
\alias{validModel}
\title{Validity Function for \linkS4class{model}}
\usage{
validModel(object)
}
\arguments{
\item{object}{Object of class \linkS4class{model}.}
}
\description{
Validity Function for object of class \linkS4class{model}.
}
