% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape-validity.R
\name{validShape}
\alias{validShape}
\title{Validity Function for \linkS4class{Shape}}
\usage{
validShape(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Shape}.}
}
\description{
Validity Function for object of class \linkS4class{Shape}.
}
