% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{verticalPlot_height}
\alias{verticalPlot_height}
\title{Get Vertical Plot Height for object of class model}
\usage{
verticalPlot_height(model, iter)
}
\arguments{
\item{model}{Object of class \linkS4class{model}.}

\item{iter}{Optional parameter indicating iteration to use for plotting. Default is latest iteration value in model.}
}
\description{
Gets vertical plot height to match output of plot(model)
}
