% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingTools.R
\name{yMarg}
\alias{yMarg}
\title{Get y-axis plotting margins}
\usage{
yMarg(x, dig = 2)
}
\arguments{
\item{x}{Value}

\item{dig}{Number of digits to use.}
}
\description{
Gets flexible y-axis margins for basic N, Shape, View, Strand, and Round plots.
}
